Generate all dependencies by bower
all dependencies will not be included in git but downloaded by bower,
this is to keep platfrom safe libraries.
in order to download all dependencies run "bower install" under assets
directory . (if you are behind proxy change the srttings in .bowerrc)
if you wnt to add new dependency , update the bower.json file.